myenv/mysite/requirements.txt
asgiref==3.3.4
Django==3.2.3
pytz==2021.1
sqlparse==0.4.1
Pillow==8.2.0
django-crispy-forms==1.11.2